third_party/blink/web_tests/external/wpt/app-history/app-history-entry/key-id-location-reload.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    let key = i.contentWindow.appHistory.current.key;
    let id = i.contentWindow.appHistory.current.id;
    i.contentWindow.location.reload();
    i.onload = t.step_func_done(() => {
      assert_equals(key, i.contentWindow.appHistory.current.key);
      assert_equals(id, i.contentWindow.appHistory.current.id);
      assert_equals(appHistory.entries().length, 1);
    });
  });
}, "AppHistoryEntry's key and id after location.reload()");
</script>
